requirements.txt
requests
beautifulsoup4
chromadb
urllib3
langchain
reportlab
pypdf